requirements.txt
# Core data processing & ML
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Deep Learning & Computer Vision
torch>=1.10.0
torchvision>=0.11.0
ultralytics>=8.0.0

# Data handling & progressbars
tqdm>=4.62.0

# Development utilities (optional)
jupyter>=1.0.0
ipython>=7.0.0

# Cloud/Kaggle (optional)
kaggle>=1.5.12
google-colab>=1.0.0

# Visualization (optional)
matplotlib>=3.4.0
